requirements.txt
Django==2.2.13
djangorestframework==3.11.0
django-filter==2.3.0
drf-yasg==1.17.1
django-cors-headers==3.4.0
boto3==1.14.12
djongo==1.3.2
sqlparse==0.2.4
djoser==2.0.3
django-cassandra-common==0.0.2